requirements.txt
praw>=7.7.0
edge-tts>=6.1.0
openai-whisper>=20231117
faster-whisper>=1.0.0
python-dotenv>=1.0.0
pydub>=0.25.1
yt-dlp>=2024.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.3
requests>=2.31.0
